requirements.txt
openai>=1.44.0
python-dotenv>=1.0.1
requests>=2.32.3
PyYAML>=6.0.2
tenacity>=9.0.0
tqdm>=4.66.5
rich>=13.8.0
jinja2>=3.1.4
pandas>=2.2.2
